requirements.txt
requests~=2.28.0
colorlog~=6.6.0
zabbix~=1.2.4
